bridge/requirements.txt
anyio==4.4.0
asyncpg==0.29.0
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jaconv==0.3.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tzdata==2024.1
uvicorn==0.30.1
